:construction: Changed EmployedUser ID type from Integer to Long

+++ src/main/java/com/schnabel/schnabel/users/repository/IEmployedUserRepository.java
@@ -3,6 +3,6 @@ package com.schnabel.schnabel.users.repository;
 import com.schnabel.schnabel.users.model.EmployedUser;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IEmployedUserRepository extends CrudRepository<EmployedUser, Integer>
+public interface IEmployedUserRepository extends CrudRepository<EmployedUser, Long>
 {
 }
